wpt_babel/content-security-policy/nonce-hiding/nonce-hiding-move-document.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<meta content="style-src 'self' 'nonce-allowme';" http-equiv="Content-Security-Policy"/>
<link href="https://bugzilla.mozilla.org/show_bug.cgi?id=1831328" rel="help"/>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<title>
 Nonce isn't lost on document move
</title>
<style nonce="allowme" type="text/css">
 p {
    color: red;
  }
</style>
<p>
 What color is this?
</p>
<script>
 "use strict";

test(function () {
  var doc = document.implementation.createDocument("http://www.w3.org/1999/xhtml", "html");
  var style = document.createElement("style");
  style.setAttribute("nonce", "allowme");
  style.textContent = "p { color: lime }";
  doc.documentElement.appendChild(style);
  document.body.appendChild(style);
  assert_equals(style.nonce, "allowme", "Nonce should not have been lost");
  assert_equals(getComputedStyle(document.querySelector("p")).color, "rgb(0, 255, 0)", "Style should apply");
});
</script>
